code migration
	1. combine hpp and cc files for templates
	2. remove files which are not usable

branch: dev
user: ysydty
