+++ .project/i18n/Creater/Creater/MacroWriter.cs
@@ -58,7 +58,7 @@ namespace DTY.Native.Tools.I18N.Creater
 
         private void _StartMacro()
         {
-            _writer.WriteLine("// ################################################");
+            _writer.WriteLine("// #####################################################");
             _writer.WriteLine("// 天宇静态多语言文件");
             _writer.WriteLine("// Dev Tianyu Static Language File");
             _writer.WriteLine("//");
@@ -70,7 +70,11 @@ namespace DTY.Native.Tools.I18N.Creater
             _writer.WriteLine("// Language   :  {0}", AreaHelper.AreaString(_area));
             _writer.WriteLine("// Copyright  :  aitianyu.cn");
             _writer.WriteLine("//");
-            _writer.WriteLine("// ################################################");
+            _writer.WriteLine("// 此文件由天宇自动化生成器生成，请勿手动更改");
+            _writer.WriteLine("// This file is generated by Tianyu Automatic Creater");
+            _writer.WriteLine("// Please do NOT manually change");
+            _writer.WriteLine("//");
+            _writer.WriteLine("// #####################################################");
             _writer.WriteLine();
 
             _writer.WriteLine("#ifndef {0}", _macroText);
