+++ native/prime/core/src/error.cpp
@@ -1,52 +1,75 @@
 #include"../error.hpp"
 #include<string>
 
-using namespace dty::except;
+#define __DTY_EXCEPT_NS__ dty::except
+
+#pragma region Exception ID Macros
+constexpr uint64 __VARIABLE__ DTY_Generic_Exception = 0x0000000000000001ULL;
+constexpr uint64 __VARIABLE__ DTY_Generic_NullPointerException = 0x0000000000000002ULL;
+constexpr uint64 __VARIABLE__ DTY_Generic_ArithmeticException = 0x0000000000000003ULL;
+constexpr uint64 __VARIABLE__ DTY_Generic_IndexOutOfRangeException = 0x0000000000000004ULL;
+constexpr uint64 __VARIABLE__ DTY_Generic_NotImplementationException = 0x0000000000000005ULL;
+
+constexpr uint64 __VARIABLE__ DTY_Generic_ArgumentException = 0x1000000000000001ULL;
+constexpr uint64 __VARIABLE__ DTY_Generic_ArgumentNullException = 0x1000000000000002ULL;
+constexpr uint64 __VARIABLE__ DTY_Generic_ArgumentOutOfRangeException = 0x1000000000000003ULL;
+
+constexpr uint64 __VARIABLE__ DTY_Generic_OperationNotSupportException = 0x2000000000000001ULL;
+
+constexpr uint64 __VARIABLE__ DTY_Generic_OperationFailedException = 0x3000000000000001ULL;
+#pragma endregion
 
 // #######################################################
 // Tianyu Exception Base
 // #######################################################
 
 #pragma region Tianyu Exception Base
-Exception::Exception()
+__DTY_EXCEPT_NS__::Exception::Exception()
     : dty::TianyuObject(),
     std::exception(),
     dty::collection::IEquatable<Exception>(),
     dty::collection::ICompareable<Exception>()
 { }
-Exception::~Exception() { }
-uint64 Exception::GetExceptionId()
+
+__DTY_EXCEPT_NS__::Exception::~Exception() { }
+
+uint64 __DTY_EXCEPT_NS__::Exception::GetExceptionId()
 {
     return DTY_Generic_Exception;
 }
-const char* Exception::what() const noexcept
+const char* __DTY_EXCEPT_NS__::Exception::what() const noexcept
 {
     return std::exception::what();
 }
-string Exception::ToString() noexcept
+
+::string __DTY_EXCEPT_NS__::Exception::ToString() noexcept
 {
     return dty::_dty_native_cpp_default_to_string(__PTR_TO_REF__ this);
 }
-uint64 Exception::GetTypeId()
+
+uint64 __DTY_EXCEPT_NS__::Exception::GetTypeId()
 {
     return dty::GetType(__PTR_TO_REF__ this).Id();
 }
-bool Exception::Equals(Exception& other)
+
+bool __DTY_EXCEPT_NS__::Exception::Equals(__DTY_EXCEPT_NS__::Exception& other)
 {
     return (__PTR_TO_REF__ this) == other;
 }
-bool Exception::operator==(Exception& other)
+
+bool __DTY_EXCEPT_NS__::Exception::operator==(__DTY_EXCEPT_NS__::Exception& other)
 {
     return this->GetExceptionId() == other.GetExceptionId();
 }
-bool Exception::operator!=(Exception& other)
+
+bool __DTY_EXCEPT_NS__::Exception::operator!=(__DTY_EXCEPT_NS__::Exception& other)
 {
     return this->GetExceptionId() != other.GetExceptionId();
 }
 
-dty::collection::CompareResult Exception::CompareTo(Exception& other)
+dty::collection::CompareResult __DTY_EXCEPT_NS__::Exception::CompareTo(__DTY_EXCEPT_NS__::Exception& other)
 {
-    Exception& _this = (__PTR_TO_REF__ this);
+    __DTY_EXCEPT_NS__::Exception& _this = (__PTR_TO_REF__ this);
 
     return _this == other
         ? dty::collection::CompareResult::EQUAL
@@ -54,19 +77,23 @@ dty::collection::CompareResult Exception::CompareTo(Exception& other)
         ? dty::collection::CompareResult::GREAT
         : dty::collection::CompareResult::LESS;
 }
-bool Exception::operator<(Exception& other)
+
+bool __DTY_EXCEPT_NS__::Exception::operator<(__DTY_EXCEPT_NS__::Exception& other)
 {
     return this->GetExceptionId() > other.GetExceptionId();
 }
-bool Exception::operator>(Exception& other)
+
+bool __DTY_EXCEPT_NS__::Exception::operator>(__DTY_EXCEPT_NS__::Exception& other)
 {
     return this->GetExceptionId() < other.GetExceptionId();
 }
-bool Exception::operator<=(Exception& other)
+
+bool __DTY_EXCEPT_NS__::Exception::operator<=(__DTY_EXCEPT_NS__::Exception& other)
 {
     return this->GetExceptionId() >= other.GetExceptionId();
 }
-bool Exception::operator>=(Exception& other)
+
+bool __DTY_EXCEPT_NS__::Exception::operator>=(__DTY_EXCEPT_NS__::Exception& other)
 {
     return this->GetExceptionId() <= other.GetExceptionId();
 }
@@ -77,20 +104,20 @@ bool Exception::operator>=(Exception& other)
 // #######################################################
 
 #pragma region Tianyu Null Pointer Exception
-NullPointerException::NullPointerException()
-    : Exception()
+__DTY_EXCEPT_NS__::NullPointerException::NullPointerException()
+    : __DTY_EXCEPT_NS__::Exception()
 {
 
 }
-NullPointerException::~NullPointerException()
+__DTY_EXCEPT_NS__::NullPointerException::~NullPointerException()
 {
 
 }
-uint64 NullPointerException::GetExceptionId()
+uint64 __DTY_EXCEPT_NS__::NullPointerException::GetExceptionId()
 {
     return DTY_Generic_NullPointerException;
 }
-uint64 NullPointerException::GetTypeId()
+uint64 __DTY_EXCEPT_NS__::NullPointerException::GetTypeId()
 {
     return dty::GetType(__PTR_TO_REF__ this).Id();
 }
@@ -101,20 +128,20 @@ uint64 NullPointerException::GetTypeId()
 // #######################################################
 
 #pragma region Tianyu Arithmetic Exception
-ArithmeticException::ArithmeticException()
-    : Exception()
+__DTY_EXCEPT_NS__::ArithmeticException::ArithmeticException()
+    : __DTY_EXCEPT_NS__::Exception()
 {
 
 }
-ArithmeticException::~ArithmeticException()
+__DTY_EXCEPT_NS__::ArithmeticException::~ArithmeticException()
 {
 
 }
-uint64 ArithmeticException::GetExceptionId()
+uint64 __DTY_EXCEPT_NS__::ArithmeticException::GetExceptionId()
 {
     return DTY_Generic_ArithmeticException;
 }
-uint64 ArithmeticException::GetTypeId()
+uint64 __DTY_EXCEPT_NS__::ArithmeticException::GetTypeId()
 {
     return dty::GetType(__PTR_TO_REF__ this).Id();
 }
@@ -125,20 +152,20 @@ uint64 ArithmeticException::GetTypeId()
 // #######################################################
 
 #pragma region Tianyu Index Out of Range Exception
-IndexOutOfRangeException::IndexOutOfRangeException()
-    : Exception()
+__DTY_EXCEPT_NS__::IndexOutOfRangeException::IndexOutOfRangeException()
+    : __DTY_EXCEPT_NS__::Exception()
 {
 
 }
-IndexOutOfRangeException::~IndexOutOfRangeException()
+__DTY_EXCEPT_NS__::IndexOutOfRangeException::~IndexOutOfRangeException()
 {
 
 }
-uint64 IndexOutOfRangeException::GetExceptionId()
+uint64 __DTY_EXCEPT_NS__::IndexOutOfRangeException::GetExceptionId()
 {
     return DTY_Generic_IndexOutOfRangeException;
 }
-uint64 IndexOutOfRangeException::GetTypeId()
+uint64 __DTY_EXCEPT_NS__::IndexOutOfRangeException::GetTypeId()
 {
     return dty::GetType(__PTR_TO_REF__ this).Id();
 }
@@ -149,20 +176,20 @@ uint64 IndexOutOfRangeException::GetTypeId()
 // #######################################################
 
 #pragma region Tianyu Argument Exception
-ArgumentException::ArgumentException()
-    : Exception()
+__DTY_EXCEPT_NS__::ArgumentException::ArgumentException()
+    : __DTY_EXCEPT_NS__::Exception()
 {
 
 }
-ArgumentException::~ArgumentException()
+__DTY_EXCEPT_NS__::ArgumentException::~ArgumentException()
 {
 
 }
-uint64 ArgumentException::GetExceptionId()
+uint64 __DTY_EXCEPT_NS__::ArgumentException::GetExceptionId()
 {
     return DTY_Generic_ArgumentException;
 }
-uint64 ArgumentException::GetTypeId()
+uint64 __DTY_EXCEPT_NS__::ArgumentException::GetTypeId()
 {
     return dty::GetType(__PTR_TO_REF__ this).Id();
 }
@@ -173,20 +200,20 @@ uint64 ArgumentException::GetTypeId()
 // #######################################################
 
 #pragma region Tianyu Argument Null Pointer Exception
-ArgumentNullException::ArgumentNullException()
-    : ArgumentException()
+__DTY_EXCEPT_NS__::ArgumentNullException::ArgumentNullException()
+    : __DTY_EXCEPT_NS__::ArgumentException()
 {
 
 }
-ArgumentNullException::~ArgumentNullException()
+__DTY_EXCEPT_NS__::ArgumentNullException::~ArgumentNullException()
 {
 
 }
-uint64 ArgumentNullException::GetExceptionId()
+uint64 __DTY_EXCEPT_NS__::ArgumentNullException::GetExceptionId()
 {
     return DTY_Generic_ArgumentNullException;
 }
-uint64 ArgumentNullException::GetTypeId()
+uint64 __DTY_EXCEPT_NS__::ArgumentNullException::GetTypeId()
 {
     return dty::GetType(__PTR_TO_REF__ this).Id();
 }
@@ -197,20 +224,20 @@ uint64 ArgumentNullException::GetTypeId()
 // #######################################################
 
 #pragma region Tianyu Argument Out of Range Exception
-ArgumentOutOfRangeException::ArgumentOutOfRangeException()
-    : ArgumentException()
+__DTY_EXCEPT_NS__::ArgumentOutOfRangeException::ArgumentOutOfRangeException()
+    : __DTY_EXCEPT_NS__::ArgumentException()
 {
 
 }
-ArgumentOutOfRangeException::~ArgumentOutOfRangeException()
+__DTY_EXCEPT_NS__::ArgumentOutOfRangeException::~ArgumentOutOfRangeException()
 {
 
 }
-uint64 ArgumentOutOfRangeException::GetExceptionId()
+uint64 __DTY_EXCEPT_NS__::ArgumentOutOfRangeException::GetExceptionId()
 {
     return DTY_Generic_ArgumentOutOfRangeException;
 }
-uint64 ArgumentOutOfRangeException::GetTypeId()
+uint64 __DTY_EXCEPT_NS__::ArgumentOutOfRangeException::GetTypeId()
 {
     return dty::GetType(__PTR_TO_REF__ this).Id();
 }
@@ -221,20 +248,20 @@ uint64 ArgumentOutOfRangeException::GetTypeId()
 // #######################################################
 
 #pragma region Tianyu Not Implementation Exception
-NotImplementationException::NotImplementationException()
-    : Exception()
+__DTY_EXCEPT_NS__::NotImplementationException::NotImplementationException()
+    : __DTY_EXCEPT_NS__::Exception()
 {
 
 }
-NotImplementationException::~NotImplementationException()
+__DTY_EXCEPT_NS__::NotImplementationException::~NotImplementationException()
 {
 
 }
-uint64 NotImplementationException::GetExceptionId()
+uint64 __DTY_EXCEPT_NS__::NotImplementationException::GetExceptionId()
 {
     return DTY_Generic_NotImplementationException;
 }
-uint64 NotImplementationException::GetTypeId()
+uint64 __DTY_EXCEPT_NS__::NotImplementationException::GetTypeId()
 {
     return dty::GetType(__PTR_TO_REF__ this).Id();
 }
@@ -245,20 +272,20 @@ uint64 NotImplementationException::GetTypeId()
 // #######################################################
 
 #pragma region Tianyu Operation Not Support Exception
-OperationNotSupportException::OperationNotSupportException()
-    : Exception()
+__DTY_EXCEPT_NS__::OperationNotSupportException::OperationNotSupportException()
+    : __DTY_EXCEPT_NS__::Exception()
 {
 
 }
-OperationNotSupportException::~OperationNotSupportException()
+__DTY_EXCEPT_NS__::OperationNotSupportException::~OperationNotSupportException()
 {
 
 }
-uint64 OperationNotSupportException::GetExceptionId()
+uint64 __DTY_EXCEPT_NS__::OperationNotSupportException::GetExceptionId()
 {
     return DTY_Generic_OperationNotSupportException;
 }
-uint64 OperationNotSupportException::GetTypeId()
+uint64 __DTY_EXCEPT_NS__::OperationNotSupportException::GetTypeId()
 {
     return dty::GetType(__PTR_TO_REF__ this).Id();
 }
@@ -269,20 +296,20 @@ uint64 OperationNotSupportException::GetTypeId()
 // #######################################################
 
 #pragma region Tianyu Operation Failed Exception
-OperationFailedException::OperationFailedException()
-    : Exception()
+__DTY_EXCEPT_NS__::OperationFailedException::OperationFailedException()
+    : __DTY_EXCEPT_NS__::Exception()
 {
 
 }
-OperationFailedException::~OperationFailedException()
+__DTY_EXCEPT_NS__::OperationFailedException::~OperationFailedException()
 {
 
 }
-uint64 OperationFailedException::GetExceptionId()
+uint64 __DTY_EXCEPT_NS__::OperationFailedException::GetExceptionId()
 {
     return DTY_Generic_OperationFailedException;
 }
-uint64 OperationFailedException::GetTypeId()
+uint64 __DTY_EXCEPT_NS__::OperationFailedException::GetTypeId()
 {
     return dty::GetType(__PTR_TO_REF__ this).Id();
 }
